backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
firebase-admin==6.5.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1

# ML
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
ortools==9.9.3963
flask==3.0.3
gunicorn==22.0.0

# Utilities
python-dateutil==2.9.0
geopy==2.4.1
shapely==2.0.4
